requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
selenium==4.15.2
webdriver-manager==4.0.1
selenium-stealth==1.0.6
fake-useragent==1.4.0
python-telegram-bot==20.7
python-dotenv==1.0.0
rich==13.7.0
jinja2==3.1.2
aiofiles==23.2.1
